pyproject.toml
[project]
name = "pair-trading"
version = "0.1.0"
description = "Copula-based pair trading for cryptocurrency futures"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "copulae>=0.7.9",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "copulae>=0.7.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "requests>=2.31.0",
]
